load_test/requirements.txt
locust==2.18.0
python-dotenv==1.0.0
kafka-python==2.0.2
